change plist path to directory instead of file
The plistreader section of the recipe expected Tableau.app, but what was really there was Tableau 2020.3.app. This would be a problem in the future as Tableau changes versions, so I changed the path to lead to the folder instead of directly to the plist. Plistreader is able to search through bundles in directories automatically so this shouldn't be a problem.